mv `ActionItem` to `ui-core`
