+++ packages/app-staking/src/modals/Accounts/index.tsx
@@ -5,10 +5,9 @@ import { useActiveAccount, useImportedAccounts } from '@polkadot-cloud/connect'
 import { useProxies } from '@polkadot-cloud/connect-proxies'
 import { useActiveProxy } from 'hooks/useActiveProxy'
 import { useBalances } from 'hooks/useBalances'
-import { ActionItem } from 'library/ActionItem'
 import { useEffect } from 'react'
 import { useTranslation } from 'react-i18next'
-import { CustomHeader, Padding } from 'ui-core/modal'
+import { ActionItem, CustomHeader, Padding } from 'ui-core/modal'
 import { Close, useOverlay } from 'ui-overlay'
 import { AccountItem } from './AccountItem'
 import type {
