+++ packages/app-staking/src/modals/LeavePool/index.tsx
@@ -15,14 +15,13 @@ import { useNetwork } from 'hooks/useNetwork'
 import { useSignerWarnings } from 'hooks/useSignerWarnings'
 import { useSubmitExtrinsic } from 'hooks/useSubmitExtrinsic'
 import { formatFromProp } from 'hooks/useSubmitExtrinsic/util'
-import { ActionItem } from 'library/ActionItem'
 import { Warning } from 'library/Form/Warning'
 import { ModalBack } from 'library/ModalBack'
 import { SubmitTx } from 'library/SubmitTx'
 import { StaticNote } from 'modals/Utils/StaticNote'
 import { useEffect, useState } from 'react'
 import { useTranslation } from 'react-i18next'
-import { Padding, Title, Warnings } from 'ui-core/modal'
+import { ActionItem, Padding, Title, Warnings } from 'ui-core/modal'
 import { useOverlay } from 'ui-overlay'
 import { timeleftAsString } from 'utils'
