+++ packages/app-staking/src/modals/ManagePool/Forms/ClaimCommission/index.tsx
@@ -11,14 +11,13 @@ import { useNetwork } from 'hooks/useNetwork'
 import { useSignerWarnings } from 'hooks/useSignerWarnings'
 import { useSubmitExtrinsic } from 'hooks/useSubmitExtrinsic'
 import { formatFromProp } from 'hooks/useSubmitExtrinsic/util'
-import { ActionItem } from 'library/ActionItem'
 import { Warning } from 'library/Form/Warning'
 import { ModalBack } from 'library/ModalBack'
 import { SubmitTx } from 'library/SubmitTx'
 import type { Dispatch, SetStateAction } from 'react'
 import { useEffect, useState } from 'react'
 import { useTranslation } from 'react-i18next'
-import { Notes, Padding, Warnings } from 'ui-core/modal'
+import { ActionItem, Notes, Padding, Warnings } from 'ui-core/modal'
 import { useOverlay } from 'ui-overlay'
 import { planckToUnitBn } from 'utils'
