+++ packages/app-staking/src/modals/ManagePool/Forms/SetPoolState/index.tsx
@@ -10,14 +10,13 @@ import { useApi } from 'hooks/useApi'
 import { useSignerWarnings } from 'hooks/useSignerWarnings'
 import { useSubmitExtrinsic } from 'hooks/useSubmitExtrinsic'
 import { formatFromProp } from 'hooks/useSubmitExtrinsic/util'
-import { ActionItem } from 'library/ActionItem'
 import { Warning } from 'library/Form/Warning'
 import { ModalBack } from 'library/ModalBack'
 import { SubmitTx } from 'library/SubmitTx'
 import type { Dispatch, SetStateAction } from 'react'
 import { useEffect, useState } from 'react'
 import { useTranslation } from 'react-i18next'
-import { Padding, Warnings } from 'ui-core/modal'
+import { ActionItem, Padding, Warnings } from 'ui-core/modal'
 import { useOverlay } from 'ui-overlay'
 
 export const SetPoolState = ({
