+++ packages/app-staking/src/modals/UnlockChunks/Forms.tsx
@@ -14,13 +14,12 @@ import { useNetwork } from 'hooks/useNetwork'
 import { useSignerWarnings } from 'hooks/useSignerWarnings'
 import { useSubmitExtrinsic } from 'hooks/useSubmitExtrinsic'
 import { formatFromProp } from 'hooks/useSubmitExtrinsic/util'
-import { ActionItem } from 'library/ActionItem'
 import { Warning } from 'library/Form/Warning'
 import { ModalBack } from 'library/ModalBack'
 import { SubmitTx } from 'library/SubmitTx'
 import { type ForwardedRef, forwardRef, useEffect, useState } from 'react'
 import { useTranslation } from 'react-i18next'
-import { Padding, Warnings } from 'ui-core/modal'
+import { ActionItem, Padding, Warnings } from 'ui-core/modal'
 import { useOverlay } from 'ui-overlay'
 import type { FormsProps } from './types'
 import { ContentWrapper } from './Wrappers'
