+++ packages/ui-core/src/modal/ActionItem/index.tsx
@@ -4,8 +4,8 @@
 import { faCheck, faChevronRight } from '@fortawesome/free-solid-svg-icons'
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
 import { useEffect, useState } from 'react'
-import type { ActionItemProps } from './types'
-import { Wrapper } from './Wrapper'
+import type { ActionItemProps } from '../types'
+import classes from './index.module.scss'
 
 /**
  * @name ActionItem
@@ -24,8 +24,10 @@ export const ActionItem = ({
 	const [toggle, setToggle] = useState<boolean | undefined>(toggled)
 
 	useEffect(() => setToggle(toggled), [toggled])
+
 	return (
-		<Wrapper
+		<h3
+			className={classes.actionItem}
 			style={{
 				...style,
 				opacity: inactive ? 0.3 : 1,
@@ -36,7 +38,7 @@ export const ActionItem = ({
 			) : (
 				<button
 					type="button"
-					className="toggle"
+					className={classes.toggle}
 					disabled={disabled}
 					onClick={() => {
 						if (typeof onToggle === 'function') {
@@ -50,6 +52,6 @@ export const ActionItem = ({
 			)}
 			{text}
 			{inlineButton && <span>{inlineButton}</span>}
-		</Wrapper>
+		</h3>
 	)
 }
